+++ src/Tesseract/InteropDotNet/UnixLibraryLoaderLogic.cs
@@ -4,6 +4,8 @@
 using System;
 using System.Runtime.InteropServices;
 
+using Tesseract.Internal;
+
 namespace InteropDotNet
 {
     class UnixLibraryLoaderLogic : ILibraryLoaderLogic
@@ -14,17 +16,17 @@ namespace InteropDotNet
 
             try
             {
-                LibraryLoaderTrace.TraceInformation("Trying to load native library \"{0}\"...", fileName);
+                Logger.TraceInformation("Trying to load native library \"{0}\"...", fileName);
                 libraryHandle = UnixLoadLibrary(fileName, RTLD_NOW);
                 if (libraryHandle != IntPtr.Zero)
-                    LibraryLoaderTrace.TraceInformation("Successfully loaded native library \"{0}\", handle = {1}.", fileName, libraryHandle);
+                    Logger.TraceInformation("Successfully loaded native library \"{0}\", handle = {1}.", fileName, libraryHandle);
                 else
-                    LibraryLoaderTrace.TraceError("Failed to load native library \"{0}\".\r\nCheck windows event log.", fileName);
+                    Logger.TraceError("Failed to load native library \"{0}\".\r\nCheck windows event log.", fileName);
             }
             catch (Exception e)
             {
                 var lastError = UnixGetLastError();
-                LibraryLoaderTrace.TraceError("Failed to load native library \"{0}\".\r\nLast Error:{1}\r\nCheck inner exception and\\or windows event log.\r\nInner Exception: {2}", fileName, lastError, e.ToString());
+                Logger.TraceError("Failed to load native library \"{0}\".\r\nLast Error:{1}\r\nCheck inner exception and\\or windows event log.\r\nInner Exception: {2}", fileName, lastError, e.ToString());
             }
 
             return libraryHandle;
@@ -38,17 +40,17 @@ namespace InteropDotNet
         public IntPtr GetProcAddress(IntPtr libraryHandle, string functionName)
         {
             UnixGetLastError(); // Clearing previous errors
-            LibraryLoaderTrace.TraceInformation("Trying to load native function \"{0}\" from the library with handle {1}...",
+            Logger.TraceInformation("Trying to load native function \"{0}\" from the library with handle {1}...",
                 functionName, libraryHandle);
             var functionHandle = UnixGetProcAddress(libraryHandle, functionName);
             var errorPointer = UnixGetLastError();
             if (errorPointer != IntPtr.Zero)
                 throw new Exception("dlsym: " + Marshal.PtrToStringAnsi(errorPointer));
             if (functionHandle != IntPtr.Zero && errorPointer == IntPtr.Zero)
-                LibraryLoaderTrace.TraceInformation("Successfully loaded native function \"{0}\", function handle = {1}.",
+                Logger.TraceInformation("Successfully loaded native function \"{0}\", function handle = {1}.",
                     functionName, functionHandle);
             else
-                LibraryLoaderTrace.TraceError("Failed to load native function \"{0}\", function handle = {1}, error pointer = {2}",
+                Logger.TraceError("Failed to load native function \"{0}\", function handle = {1}, error pointer = {2}",
                     functionName, functionHandle, errorPointer);
             return functionHandle;
         }