+++ src/Tesseract/TesseractEngine.cs
@@ -14,7 +14,7 @@ namespace Tesseract
     /// </summary>
     public class TesseractEngine : DisposableBase
     {
-        private const string TesseractVersion = "3.04.01";
+        private const string TesseractVersion = "3.05.02";
         private static readonly TraceSource trace = new TraceSource("Tesseract");
 
         private HandleRef handle;